refactor(sortIcon): move sortIcon to a new file
